Need to support make and make install stages

+++ Code/JavaWrappers/csharp_wrapper/fix_msvc_swig.py
@@ -2,6 +2,7 @@
 import sys
 
 for file in sys.argv[1:]:
-    text = open(file).read()
-    text = text.replace("|| (LONG_MAX == INT_MAX)", "")
-    open(file,'w').write(text)
+    if os.path.exists(file):
+        text = open(file).read()
+        text = text.replace("|| (LONG_MAX == INT_MAX)", "")
+        open(file,'w').write(text)
